+++ packages/cloud-cognitive/src/components/TagSet/TagSet.stories.js
@@ -7,7 +7,6 @@
 
 import React from 'react';
 
-import { Tag } from 'carbon-components-react';
 import { types as tagTypes } from 'carbon-components-react/es/components/Tag/Tag';
 import styles from './_storybook-styles.scss'; // import index in case more files are added later.
 import { pkg } from '../../settings';
@@ -19,27 +18,18 @@ const storybookPrefix = getStorybookPrefix(pkg, TagSet.displayName);
 const blockClass = `${pkg.prefix}--tag-set`;
 const blockClassModal = `${blockClass}-modal`;
 
-const TagItems = [
-  <Tag key="tag1" type="blue">
-    Tag 1
-  </Tag>,
-  <Tag key="tag12">Tag 12</Tag>,
-  <Tag key="tag123" type="high-contrast">
-    Tag 123
-  </Tag>,
-  <Tag key="tag1234" type="cyan">
-    Tag 1234
-  </Tag>,
-  <Tag key="tag12345" type="red">
-    Tag 12345
-  </Tag>,
+const tags = [
+  { type: 'blue', label: 'Tag 1' },
+  { type: 'high-contrast', label: 'Tag 123' },
+  { type: 'cyan', label: 'Tag 1234' },
+  { type: 'red', label: 'Tag 12345' },
 ];
 
-const ManyTagItems = [
+const manyTags = [
   {
     label: 'One',
     type: 'blue',
-    dataSearch: 'one',
+    ['data-search']: 'single',
   },
   {
     label: 'Two',
@@ -84,7 +74,7 @@ const ManyTagItems = [
   {
     label: 'Twelve',
     type: 'high-contrast',
-    dataSearch: 'twelve',
+    ['data-search']: 'dozen',
   },
   {
     label: 'Thirteen',
@@ -118,22 +108,17 @@ const ManyTagItems = [
     label: 'Twenty',
     type: 'high-contrast',
   },
-].map(({ label, type, dataSearch }) => (
-  <Tag key={label} data-search={dataSearch} type={type}>
-    {label}
-  </Tag>
-));
+].map((item, index) => ({
+  ...item,
+  ['data-search']: '' + (index + 1) + ' ' + item?.['data-search'],
+}));
 
-const HundredsOfItems = [];
+const hundredsOfTags = [];
 for (let i = 0; i < 200; i++) {
   const label = `Label_${i + 1}`;
   const type = tagTypes[i % tagTypes.length];
 
-  HundredsOfItems.push(
-    <Tag key={label} type={type}>
-      {label}
-    </Tag>
-  );
+  hundredsOfTags.push({ type, label });
 }
 
 const overflowAndModalStrings = {
@@ -169,30 +154,30 @@ export default {
 };
 
 const Template = (argsIn) => {
-  const { children, containerWidth, ...args } = { ...argsIn };
+  const { containerWidth, ...args } = { ...argsIn };
   return (
     <div style={{ width: containerWidth }}>
-      <TagSet {...args}>{children}</TagSet>
+      <TagSet {...args} />
     </div>
   );
 };
 
 export const FiveTags = Template.bind({});
 FiveTags.args = {
-  children: TagItems,
+  tags: tags,
   containerWidth: 500,
 };
 
 export const ManyTags = Template.bind({});
 ManyTags.args = {
-  children: ManyTagItems,
+  tags: manyTags,
   containerWidth: 500,
   ...overflowAndModalStrings,
 };
 
 export const HundredsOfTags = Template.bind({});
 HundredsOfTags.args = {
-  children: HundredsOfItems,
+  tags: hundredsOfTags,
   containerWidth: 500,
   ...overflowAndModalStrings,
 };