+++ packages/cloud-cognitive/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.47.7](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.47.6...@carbon/ibm-cloud-cognitive@0.47.7) (2021-06-22)
+
+
+### Bug Fixes
+
+* **CreateTearsheet:** remove overflow styles ([#924](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/924)) ([369329a](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/369329a34cd3ea45ae8e4c21938d358a78fd92cb))
+
+
+
+
+
+## [0.47.6](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.47.5...@carbon/ibm-cloud-cognitive@0.47.6) (2021-06-22)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
+## [0.47.5](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.47.4...@carbon/ibm-cloud-cognitive@0.47.5) (2021-06-22)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
+## [0.47.4](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.47.3...@carbon/ibm-cloud-cognitive@0.47.4) (2021-06-21)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
 ## [0.47.3](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.47.2...@carbon/ibm-cloud-cognitive@0.47.3) (2021-06-21)
 
 **Note:** Version bump only for package @carbon/ibm-cloud-cognitive